app/api-hosted-models/assemblyai-transcription/requirements.txt
fastapi==0.95.0
pydantic==1.10.6
uvicorn[standard]
requests==2.28.2
python-dotenv==1.0.0